Abstract out common logic between EditCommand and EditUserCommand into an abstract class EditPersonCommand.
EditCommand now renamed to EditContactCommand to avoid confusion.
