+++ src/main/java/seedu/address/logic/parser/DeleteCommandParser.java
@@ -1,10 +1,13 @@
 package seedu.address.logic.parser;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
+import static seedu.address.logic.parser.ParserUtil.MESSAGE_INVALID_INDEX;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.DeleteCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
+
+import java.util.Optional;
 
 /**
  * Parses input arguments and creates a new DeleteCommand object
@@ -18,8 +21,11 @@ public class DeleteCommandParser implements Parser<DeleteCommand> {
      */
     public DeleteCommand parse(String args) throws ParseException {
         try {
-            Index index = ParserUtil.parseIndex(args);
-            return new DeleteCommand(index);
+            Optional<Index> index = ParserUtil.parseIndex(args);
+            if (index.isEmpty()) {
+                throw new ParseException(MESSAGE_INVALID_INDEX);
+            }
+            return new DeleteCommand(index.get());
         } catch (ParseException pe) {
             throw new ParseException(
                     String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteCommand.MESSAGE_USAGE), pe);
