+++ src/main/java/seedu/address/logic/parser/FavoriteCommandParser.java
@@ -2,10 +2,13 @@ package seedu.address.logic.parser;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
+import static seedu.address.logic.parser.ParserUtil.MESSAGE_INVALID_INDEX;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.FavoriteCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
+
+import java.util.Optional;
 
 /**
  * Parses input arguments and creates a new FavoriteCommand object
@@ -18,11 +21,14 @@ public class FavoriteCommandParser implements Parser<FavoriteCommand> {
      */
     public FavoriteCommand parse(String args) throws ParseException {
         requireNonNull(args);
-        Index index;
+        Optional<Index> index;
 
         try {
             index = ParserUtil.parseIndex(args);
-            return new FavoriteCommand(index);
+            if (index.isEmpty()) {
+                throw new ParseException(MESSAGE_INVALID_INDEX);
+            }
+            return new FavoriteCommand(index.get());
         } catch (ParseException pe) {
             throw new ParseException(
                     String.format(MESSAGE_INVALID_COMMAND_FORMAT, FavoriteCommand.MESSAGE_USAGE), pe);
