+++ src/main/java/seedu/address/logic/parser/UnfavoriteCommandParser.java
@@ -2,10 +2,13 @@ package seedu.address.logic.parser;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
+import static seedu.address.logic.parser.ParserUtil.MESSAGE_INVALID_INDEX;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.UnfavoriteCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
+
+import java.util.Optional;
 
 /**
  * Parses input arguments and creates a new UnfavoriteCommand object
@@ -18,11 +21,14 @@ public class UnfavoriteCommandParser implements Parser<UnfavoriteCommand> {
      */
     public UnfavoriteCommand parse(String args) throws ParseException {
         requireNonNull(args);
-        Index index;
+        Optional<Index> index;
 
         try {
             index = ParserUtil.parseIndex(args);
-            return new UnfavoriteCommand(index);
+            if (index.isEmpty()) {
+                throw new ParseException(MESSAGE_INVALID_INDEX);
+            }
+            return new UnfavoriteCommand(index.get());
         } catch (ParseException pe) {
             throw new ParseException(
                     String.format(MESSAGE_INVALID_COMMAND_FORMAT, UnfavoriteCommand.MESSAGE_USAGE), pe);
