+++ src/main/java/seedu/address/logic/parser/editpersoncommands/EditPersonCommandParser.java
@@ -1,6 +1,7 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.editpersoncommands;
 
 import static java.util.Objects.requireNonNull;
+import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_ADDRESS;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_COMMS;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_EMAIL;
@@ -17,30 +18,43 @@ import java.util.Collections;
 import java.util.Optional;
 import java.util.Set;
 
-import seedu.address.logic.commands.EditCommand;
-import seedu.address.logic.commands.EditUserCommand;
+import seedu.address.commons.core.index.Index;
+import seedu.address.logic.commands.EditContactCommand;
+import seedu.address.logic.parser.ArgumentMultimap;
+import seedu.address.logic.parser.ArgumentTokenizer;
+import seedu.address.logic.parser.ParserUtil;
 import seedu.address.logic.parser.exceptions.ParseException;
+import seedu.address.model.person.Person;
 import seedu.address.model.person.fields.Modules;
 import seedu.address.model.tag.Tag;
 
 /**
- * Parses input arguments and creates a new EditCommand object
+ * Abstract class to inherit from for parser classes which parse objects of the {@link Person} class
  */
-public class EditUserCommandParser implements Parser<EditUserCommand> {
+public class EditPersonCommandParser {
 
     /**
      * Parses the given {@code String} of arguments in the context of the EditCommand
-     * and returns an EditCommand object for execution.
+     * and returns an EditPersonDescriptor for edit commands to use.
      * @throws ParseException if the user input does not conform the expected format
      */
-    public EditUserCommand parse(String args) throws ParseException {
+    public EditPersonDescriptor parseForTags(String args) throws ParseException {
         requireNonNull(args);
         ArgumentMultimap argMultimap =
                 ArgumentTokenizer.tokenize(args,
                         PREFIX_NAME, PREFIX_PHONE, PREFIX_EMAIL, PREFIX_ADDRESS, PREFIX_TAG,
                         PREFIX_GENDER, PREFIX_MAJOR, PREFIX_MODULES, PREFIX_RACE, PREFIX_COMMS);
 
-        EditUserCommand.EditUserDescriptor editPersonDescriptor = new EditUserCommand.EditUserDescriptor();
+        Optional<Index> index;
+        EditPersonDescriptor editPersonDescriptor = new EditPersonDescriptor();
+
+        try {
+            index = ParserUtil.parseIndex(argMultimap.getPreamble());
+            editPersonDescriptor.setIndex(index);
+        } catch (ParseException pe) {
+            throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, EditContactCommand.MESSAGE_USAGE), pe);
+        }
+
         if (argMultimap.getValue(PREFIX_NAME).isPresent()) {
             editPersonDescriptor.setName(ParserUtil.parseName(argMultimap.getValue(PREFIX_NAME).get()));
         }
@@ -71,10 +85,10 @@ public class EditUserCommandParser implements Parser<EditUserCommand> {
 
 
         if (!editPersonDescriptor.isAnyFieldEdited()) {
-            throw new ParseException(EditCommand.MESSAGE_NOT_EDITED);
+            throw new ParseException(EditContactCommand.MESSAGE_NOT_EDITED);
         }
 
-        return new EditUserCommand(editPersonDescriptor);
+        return editPersonDescriptor;
     }
 
     private Optional<Modules> parseModulesForEdit(Collection<String> mods) throws ParseException {
@@ -103,3 +117,4 @@ public class EditUserCommandParser implements Parser<EditUserCommand> {
     }
 
 }
+
