+++ src/main/java/seedu/address/ui/result/ResultDisplay.java
@@ -14,7 +14,7 @@ import javafx.scene.layout.VBox;
 import seedu.address.logic.commands.AddCommand;
 import seedu.address.logic.commands.ClearCommand;
 import seedu.address.logic.commands.DeleteCommand;
-import seedu.address.logic.commands.EditCommand;
+import seedu.address.logic.commands.EditContactCommand;
 import seedu.address.logic.commands.ExitCommand;
 import seedu.address.logic.commands.FindCommand;
 import seedu.address.logic.commands.HelpCommand;
@@ -28,7 +28,7 @@ public class ResultDisplay extends UiPart<Region> {
 
     private static final String FXML = "result/ResultDisplay.fxml";
     private static final List<String> KEYWORDS = List.of(AddCommand.COMMAND_WORD, ClearCommand.COMMAND_WORD,
-            DeleteCommand.COMMAND_WORD, EditCommand.COMMAND_WORD, ExitCommand.COMMAND_WORD,
+            DeleteCommand.COMMAND_WORD, EditContactCommand.COMMAND_WORD, ExitCommand.COMMAND_WORD,
             FindCommand.COMMAND_WORD, HelpCommand.COMMAND_WORD, ListCommand.COMMAND_WORD,
             "Parameters", "Example");
